SHRINKWRAP-366 (Prototype) Add support for Workspace resolving
If a directory is found on Classpath(target/classes) and the directory contain a pom.xml two levels up, parse the pom to see if it matches one of the resolved artifacts. If it does, import the output folder instead of resolving it from the local/remote repository.

This can be done by using Aether's WorkSpace API.

note: a little hack is needed when the resolve output is resolveAsFile. Import, then Export the artifact to a temp file then return.